Fix x64 warnings

+++ src/core/Common.cpp
@@ -289,14 +289,14 @@ const UnicodeString LocalInvalidChars = L"/\\:*?\"<>|";
 //---------------------------------------------------------------------------
 UnicodeString ReplaceChar(UnicodeString Str, wchar_t A, wchar_t B)
 {
-  for (Integer Index = 0; Index < Str.Length(); Index++)
+  for (intptr_t Index = 0; Index < Str.Length(); Index++)
     if (Str[Index+1] == A) Str[Index+1] = B;
   return Str;
 }
 //---------------------------------------------------------------------------
 UnicodeString DeleteChar(UnicodeString Str, wchar_t C)
 {
-  int P;
+  intptr_t P;
   while ((P = Str.Pos(C)) > 0)
   {
     Str.Delete(P, 1);
@@ -391,7 +391,7 @@ UnicodeString DefaultStr(const UnicodeString & Str, const UnicodeString & Defaul
 //---------------------------------------------------------------------------
 UnicodeString CutToChar(UnicodeString &Str, wchar_t Ch, bool Trim)
 {
-  Integer P = Str.Pos(Ch);
+  intptr_t P = Str.Pos(Ch);
   UnicodeString Result;
   if (P)
   {
@@ -668,7 +668,7 @@ void __fastcall SplitCommand(UnicodeString Command, UnicodeString &Program,
   if (!Command.IsEmpty() && (Command[1] == L'\"'))
   {
     Command.Delete(1, 1);
-    int P = Command.Pos(L'"');
+    intptr_t P = Command.Pos(L'"');
     if (P)
     {
       Program = Command.SubString(1, P-1).Trim();
@@ -982,8 +982,8 @@ unsigned char __fastcall HexToByte(const UnicodeString Hex)
 {
   static UnicodeString Digits = L"0123456789ABCDEF";
   assert(Hex.Length() == 2);
-  int P1 = Digits.Pos((wchar_t)toupper(Hex[1]));
-  int P2 = Digits.Pos((wchar_t)toupper(Hex[2]));
+  intptr_t P1 = Digits.Pos((wchar_t)toupper(Hex[1]));
+  intptr_t P2 = Digits.Pos((wchar_t)toupper(Hex[2]));
 
   return
     static_cast<unsigned char>(((P1 <= 0) || (P2 <= 0)) ? 0 : (((P1 - 1) << 4) + (P2 - 1)));
